+++ packages/plugin-hint/src/editor.tsx
@@ -1,18 +1,9 @@
+import { StatefulPluginEditorProps } from '@edtr-io/core'
 import { ExpandableBox } from '@edtr-io/ui'
 import * as React from 'react'
-import { StatefulPluginEditorProps } from 'plugin-input-exercise/dist/core/src'
+
 import { hintState } from '.'
 
-export class HintEditor extends React.Component<
-  StatefulPluginEditorProps<typeof hintState>
-> {
-  public render(): React.ReactNode {
-    return (
-      <ExpandableBox
-        state={this.props.state}
-        editable={this.props.editable}
-        kind="Hinweis"
-      />
-    )
-  }
+export function HintEditor(props: StatefulPluginEditorProps<typeof hintState>) {
+  return <ExpandableBox {...props} kind="Hinweis" />
 }